+++ test/MicroElements.Swashbuckle.FluentValidation.Tests/UnitTest1.cs
@@ -41,7 +41,7 @@ namespace MicroElements.Swashbuckle.FluentValidation.Tests
 
         private SchemaFilterContext FilterContextFor(Type type)
         {
-            return new SchemaFilterContext(new ApiModel(type), null, null);
+            return new SchemaFilterContext(type, null, null);
         }
 
         private FluentValidationRules Subject()